Separate entity loading from map
